Split MetaDataStoreTests into smaller tests per repo
