+++ test/MLOps.NET.Azure.IntegrationTests/AzureLifeCycleCatalogTests.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using MLOps.NET.IntegrationTests;
+using System.Threading.Tasks;
+
+namespace MLOps.NET.Azure.IntegrationTests
+{
+    [TestClass]
+    [TestCategory("IntegrationTestCosmosDb")]
+    public class AzureLifeCycleCatalogTests : LifeCycleCatalogTests
+    {
+        [TestInitialize]
+        public void Initialize()
+        {
+            sut = IntegrationTestSetup.Initialize();
+        }
+
+        [TestCleanup]
+        public async Task TearDown()
+        {
+            var context = IntegrationTestSetup.CreateDbContext();
+
+            await base.TearDown(context);
+        }
+    }
+}
