+++ test/MLOps.NET.Azure.IntegrationTests/IntegrationTestSetup.cs
@@ -1,34 +1,28 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MLOps.NET.Azure.IntegrationTests.Constants;
 using MLOps.NET.Azure.Storage;
-using MLOps.NET.IntegrationTests;
 using MLOps.NET.Storage;
 using MLOps.NET.Storage.Database;
+using MLOps.NET.Storage.Interfaces;
 using MLOps.NET.Tests.Common.Configuration;
 using Moq;
-using System.Threading.Tasks;
 
 namespace MLOps.NET.Azure.IntegrationTests
 {
-    [TestClass]
-    [TestCategory("IntegrationTestCosmosDb")]
-    public class AzureMetaDataStoreTests : MetaDataStoreTests
+    internal class IntegrationTestSetup
     {
-        [TestInitialize]
-        public void Initialize()
+        internal static IMLOpsContext Initialize()
         {
             var configuration = ConfigurationFactory.GetConfiguration();
 
-            sut = new MLOpsBuilder()
+            return new MLOpsBuilder()
                 .UseModelRepository(new Mock<IModelRepository>().Object)
                 .UseCosmosDb(configuration[ConfigurationKeys.CosmosEndPoint],
                 configuration[ConfigurationKeys.CosmosAccountKey])
                 .Build();
         }
 
-        [TestCleanup]
-        public async Task TearDown()
+        internal static IMLOpsDbContext CreateDbContext()
         {
             var configuration = ConfigurationFactory.GetConfiguration();
 
@@ -40,7 +34,7 @@ namespace MLOps.NET.Azure.IntegrationTests
             var contextFactory = new DbContextFactory(options, CosmosEntityConfigurator.OnModelCreating);
             var context = contextFactory.CreateDbContext();
 
-            await base.TearDown(contextFactory.CreateDbContext());
+            return contextFactory.CreateDbContext();
         }
     }
 }
