+++ test/MLOps.NET.SQLite.IntegrationTests/SQLiteLifeCycleCatalogTests.cs
@@ -1,0 +1,25 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using MLOps.NET.IntegrationTests;
+using System.Threading.Tasks;
+
+namespace MLOps.NET.SQLite.IntegrationTests
+{
+    [TestCategory("Integration")]
+    [TestClass]
+    public class SQLiteLifeCycleCatalogTests : LifeCycleCatalogTests
+    {
+        [TestInitialize]
+        public void Initialize()
+        {
+            sut = IntegrationTestSetup.Initialize();
+        }
+
+        [TestCleanup]
+        public async Task TearDown()
+        {
+            var context = IntegrationTestSetup.CreateDbContext();
+
+            await base.TearDown(context);
+        }
+    }
+}
